improve simulator tab layout
