+++ src/urh/ui/ui_simulator_dialog.py
@@ -15,77 +15,6 @@ class Ui_DialogSimulator(object):
         self.tabWidgetSimulatorSettings = QtWidgets.QTabWidget(DialogSimulator)
         self.tabWidgetSimulatorSettings.setStyleSheet("QTabWidget::pane { border: 0; }")
         self.tabWidgetSimulatorSettings.setObjectName("tabWidgetSimulatorSettings")
-        self.tabGeneral = QtWidgets.QWidget()
-        self.tabGeneral.setObjectName("tabGeneral")
-        self.verticalLayout = QtWidgets.QVBoxLayout(self.tabGeneral)
-        self.verticalLayout.setContentsMargins(0, 0, 0, 0)
-        self.verticalLayout.setObjectName("verticalLayout")
-        self.scrollArea = QtWidgets.QScrollArea(self.tabGeneral)
-        self.scrollArea.setFrameShape(QtWidgets.QFrame.NoFrame)
-        self.scrollArea.setWidgetResizable(True)
-        self.scrollArea.setObjectName("scrollArea")
-        self.scrollAreaWidgetContents = QtWidgets.QWidget()
-        self.scrollAreaWidgetContents.setGeometry(QtCore.QRect(0, 0, 1066, 767))
-        self.scrollAreaWidgetContents.setObjectName("scrollAreaWidgetContents")
-        self.verticalLayout_2 = QtWidgets.QVBoxLayout(self.scrollAreaWidgetContents)
-        self.verticalLayout_2.setObjectName("verticalLayout_2")
-        self.label_5 = QtWidgets.QLabel(self.scrollAreaWidgetContents)
-        self.label_5.setObjectName("label_5")
-        self.verticalLayout_2.addWidget(self.label_5)
-        self.listViewSimulate = QtWidgets.QListView(self.scrollAreaWidgetContents)
-        self.listViewSimulate.setAlternatingRowColors(True)
-        self.listViewSimulate.setObjectName("listViewSimulate")
-        self.verticalLayout_2.addWidget(self.listViewSimulate)
-        self.gridLayout = QtWidgets.QGridLayout()
-        self.gridLayout.setObjectName("gridLayout")
-        self.label_4 = QtWidgets.QLabel(self.scrollAreaWidgetContents)
-        self.label_4.setObjectName("label_4")
-        self.gridLayout.addWidget(self.label_4, 0, 0, 1, 1)
-        self.spinBoxNRepeat = QtWidgets.QSpinBox(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Expanding, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.spinBoxNRepeat.sizePolicy().hasHeightForWidth())
-        self.spinBoxNRepeat.setSizePolicy(sizePolicy)
-        self.spinBoxNRepeat.setObjectName("spinBoxNRepeat")
-        self.gridLayout.addWidget(self.spinBoxNRepeat, 0, 1, 1, 1)
-        self.label_3 = QtWidgets.QLabel(self.scrollAreaWidgetContents)
-        self.label_3.setObjectName("label_3")
-        self.gridLayout.addWidget(self.label_3, 1, 0, 1, 1)
-        self.spinBoxTimeout = QtWidgets.QSpinBox(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Expanding, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.spinBoxTimeout.sizePolicy().hasHeightForWidth())
-        self.spinBoxTimeout.setSizePolicy(sizePolicy)
-        self.spinBoxTimeout.setObjectName("spinBoxTimeout")
-        self.gridLayout.addWidget(self.spinBoxTimeout, 1, 1, 1, 1)
-        self.label_7 = QtWidgets.QLabel(self.scrollAreaWidgetContents)
-        self.label_7.setObjectName("label_7")
-        self.gridLayout.addWidget(self.label_7, 2, 0, 1, 1)
-        self.comboBoxError = QtWidgets.QComboBox(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Expanding, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.comboBoxError.sizePolicy().hasHeightForWidth())
-        self.comboBoxError.setSizePolicy(sizePolicy)
-        self.comboBoxError.setObjectName("comboBoxError")
-        self.comboBoxError.addItem("")
-        self.comboBoxError.addItem("")
-        self.comboBoxError.addItem("")
-        self.gridLayout.addWidget(self.comboBoxError, 2, 1, 1, 1)
-        self.label_6 = QtWidgets.QLabel(self.scrollAreaWidgetContents)
-        self.label_6.setObjectName("label_6")
-        self.gridLayout.addWidget(self.label_6, 3, 0, 1, 1)
-        self.spinBoxRetries = QtWidgets.QSpinBox(self.scrollAreaWidgetContents)
-        self.spinBoxRetries.setMinimum(1)
-        self.spinBoxRetries.setProperty("value", 10)
-        self.spinBoxRetries.setObjectName("spinBoxRetries")
-        self.gridLayout.addWidget(self.spinBoxRetries, 3, 1, 1, 1)
-        self.verticalLayout_2.addLayout(self.gridLayout)
-        self.scrollArea.setWidget(self.scrollAreaWidgetContents)
-        self.verticalLayout.addWidget(self.scrollArea)
-        self.tabWidgetSimulatorSettings.addTab(self.tabGeneral, "")
         self.tabLog = QtWidgets.QWidget()
         self.tabLog.setObjectName("tabLog")
         self.verticalLayout_3 = QtWidgets.QVBoxLayout(self.tabLog)
@@ -275,7 +204,7 @@ class Ui_DialogSimulator(object):
         self.verticalLayout_4.addWidget(self.tabWidgetSimulatorSettings)
 
         self.retranslateUi(DialogSimulator)
-        self.tabWidgetSimulatorSettings.setCurrentIndex(4)
+        self.tabWidgetSimulatorSettings.setCurrentIndex(3)
         self.tabWidget.setCurrentIndex(0)
         self.groupBoxSimulationStatus.toggled['bool'].connect(self.frame.setVisible)
         self.groupBoxRXStatus.toggled['bool'].connect(self.frame_2.setVisible)
@@ -284,16 +213,6 @@ class Ui_DialogSimulator(object):
     def retranslateUi(self, DialogSimulator):
         _translate = QtCore.QCoreApplication.translate
         DialogSimulator.setWindowTitle(_translate("DialogSimulator", "Simulation"))
-        self.label_5.setText(_translate("DialogSimulator", "Simulate:"))
-        self.label_4.setText(_translate("DialogSimulator", "Repeat:"))
-        self.spinBoxNRepeat.setSpecialValueText(_translate("DialogSimulator", "Infinite"))
-        self.label_3.setText(_translate("DialogSimulator", "Timeout (in seconds):"))
-        self.label_7.setText(_translate("DialogSimulator", "In case of an overdue response:"))
-        self.comboBoxError.setItemText(0, _translate("DialogSimulator", "Resend last message"))
-        self.comboBoxError.setItemText(1, _translate("DialogSimulator", "Stop simulation"))
-        self.comboBoxError.setItemText(2, _translate("DialogSimulator", "Restart simulation"))
-        self.label_6.setText(_translate("DialogSimulator", "Max retries:"))
-        self.tabWidgetSimulatorSettings.setTabText(self.tabWidgetSimulatorSettings.indexOf(self.tabGeneral), _translate("DialogSimulator", "General settings"))
         self.btnLogAll.setText(_translate("DialogSimulator", "Log all"))
         self.btnLogNone.setText(_translate("DialogSimulator", "Log none"))
         self.btnToggleLog.setText(_translate("DialogSimulator", "Toggle selected"))
